Update slides and exercises
