+++ exercise-instructions.md
@@ -62,7 +62,7 @@ We have also reserved some GPU nodes for the course.
 In order to use these dedicated nodes, you need to run your job with the option `--reservation=gpu-course-day1`, such as
 
 ```shell
-srun --reservation=gpu-course-day1 --account=project_462001376 --partition=small-g --time=00:05:00 --nodes=1 --ntasks-per-node=1 --cpus-per-task=1 --gpus-per-task=1 ./executable
+srun --reservation=gpu-course-day1 --account=project_462001376 --partition=small-g --time=00:05:00 --nodes=1 --ntasks-per-node=1 --cpus-per-task=1 --gpus-per-node=1 ./executable
 ```
 
 On Monday, and using the reservation `gpu-course-day2` on Tuesday and `gpu-course-day3` on Wednesday.
@@ -73,11 +73,13 @@ For a multi-gpu application more cards can be requested.
 
 The common part for all of these examples includes: `srun --account=project_462001376 --partition=small-g --time=00:05:00`
 
-- 1 MPI process(es), 1 GPU(s) per process, 1 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=1 --cpus-per-task=1 --gpus-per-task=1`
-- 1 MPI process(es), 3 GPU(s) per process, 1 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=1 --cpus-per-task=1 --gpus-per-task=3`
-- 3 MPI process(es), 1 GPU(s) per process, 1 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=3 --cpus-per-task=1 --gpus-per-task=1`
-- 3 MPI process(es), 1 GPU(s) per process, 7 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=3 --cpus-per-task=7 --gpus-per-task=1`
-- 2 MPI process(es), 3 GPU(s) per process, 7 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=2 --cpus-per-task=7 --gpus-per-task=3`
+- 1 MPI process(es), 1 GPU(s) per process, 1 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=1 --cpus-per-task=1 --gpus-per-node=1`
+- 1 MPI process(es), 3 GPU(s) per process, 1 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=1 --cpus-per-task=1 --gpus-per-node=3`
+- 3 MPI process(es), 1 GPU(s) per process, 1 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=3 --cpus-per-task=1 --gpus-per-node=1`
+- 3 MPI process(es), 1 GPU(s) per process, 7 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=3 --cpus-per-task=7 --gpus-per-node=1`
+- 2 MPI process(es), 3 GPU(s) per process, 7 OpenMP thread(s) per process: `--nodes=1 --ntasks-per-node=2 --cpus-per-task=7 --gpus-per-nodes=3`
+
+See the page for example Slurm scripts on LUMI: https://docs.lumi-supercomputer.eu/runjobs/scheduled-jobs/lumig-job/
 
 ## Mahti
 In case LUMI is inaccessible during the training, we can use Mahti as a backup.
