+++ memory/03-mempools/solution/mempools.cpp
@@ -3,12 +3,6 @@
 #include <time.h>
 #include <hip/hip_runtime.h>
 
-// Undocumented extra: Enable Umpire memory management library (https://github.com/LLNL/Umpire)
-// Install Umpire and enable this if you are adventurous, no support guaranteed.
-#if defined(HAVE_UMPIRE)
-  #include "umpire/interface/c_fortran/umpire.h"
-#endif
-
 /* Blocksize divisible by the warp size */
 #define BLOCKSIZE 64
 
@@ -143,41 +137,6 @@ void recurringAllocMallocAsync(int nSteps, int size)
   HIP_ERRCHK(hipStreamDestroy(stream));
 }
 
-#if defined(HAVE_UMPIRE)
-/* Do recurring allocation with Umpire memory pool */
-void recurringAllocUmpire(int nSteps, int size)
-{
-  // Get Umpire pinned device memory pool
-  umpire_resourcemanager rm;
-  umpire_resourcemanager_get_instance(&rm);
-  umpire_allocator allocator;
-  umpire_resourcemanager_get_allocator_by_name(&rm, "DEVICE", &allocator);
-  umpire_allocator pool;
-  umpire_resourcemanager_make_allocator_quick_pool(&rm, "pool", allocator, 1024, 1024, &pool);
-
-  // Determine grid and block size
-  const int blocksize = BLOCKSIZE;
-  const int gridsize = (size - 1 + blocksize) / blocksize;
-
-  // Start timer and begin stepping loop
-  clock_t tStart = clock();
-  for(unsigned int i = 0; i < nSteps; i++)
-  {
-    int *d_A;
-    // Allocate pinned device memory with Umpire
-    d_A = (int*) umpire_allocator_allocate(&pool, sizeof(int) * size);
-    // Launch GPU kernel
-    hipKernel<<<gridsize, blocksize, 0, 0>>>(d_A, size);
-    // Free Umpire allocation
-    umpire_allocator_deallocate(&pool, d_A);
-  }
-  // Synchronization
-  HIP_ERRCHK(hipStreamSynchronize(0));
-  // Check results and print timings
-  checkTiming("recurringAllocUmpire", (double)(clock() - tStart) / CLOCKS_PER_SEC);
-}
-#endif
-
 /* The main function */
 int main(int argc, char* argv[])
 {
@@ -191,7 +150,4 @@ int main(int argc, char* argv[])
   noRecurringAlloc(nSteps, size);
   recurringAllocNoMemPools(nSteps, size);
   recurringAllocMallocAsync(nSteps, size);
-  #if defined(HAVE_UMPIRE)
-    recurringAllocUmpire(nSteps, size);
-  #endif
 }
